Data binding home + room kos

+++ app/src/main/java/com/example/b_160419073_projectuts/view/DetailKosFragment.kt
@@ -32,7 +32,7 @@ class DetailKosFragment : Fragment() {
 
         viewModel = ViewModelProvider(this).get(DetilVM::class.java)
 
-        viewModel.detail(id)
+        viewModel.detail(id.toInt())
 
         observeViewModel()
     }
